move benchmark-related code to the example directory.
