+++ tests/test_infer/test_ops/triton/test_rmsnorm_triton.py
@@ -1,6 +1,5 @@
 import pytest
 import torch
-import triton
 from packaging import version
 from transformers.models.llama.modeling_llama import LlamaRMSNorm
 
@@ -52,70 +51,5 @@ def test_layer_norm(M, N):
     assert torch.allclose(x, residual, atol=1e-5, rtol=1e-3)
 
 
-# Triton benchmark plot attributions
-configs = [
-    triton.testing.Benchmark(
-        x_names=["SEQUENCE_TOTAL"],
-        x_vals=[i for i in range(128, 1025, 128)],
-        line_arg="provider",
-        line_vals=[
-            "vllm_rms_layernorm",
-            "triton_rms_layernorm",
-            "triton_rms_layernorm_with_residual",
-            "vllm_rms_layernorm_with_residual",
-        ],
-        line_names=[
-            "vllm_rms_layernorm",
-            "triton_rms_layernorm",
-            "triton_rms_layernorm_with_residual",
-            "vllm_rms_layernorm_with_residual",
-        ],
-        styles=[("red", "-"), ("blue", "-"), ("yellow", "-"), ("green", "-")],
-        ylabel="ms",
-        plot_name=f"RMSNorm benchmarking results",
-        args={"HIDDEN_SIZE": 1024},
-    )
-]
-
-
-@triton.testing.perf_report(configs)
-def benchmark_rms_layernorm(
-    provider: str,
-    SEQUENCE_TOTAL: int,
-    HIDDEN_SIZE: int,
-):
-    try:
-        from vllm.model_executor.layers.layernorm import RMSNorm
-    except ImportError:
-        raise ImportError("Please install vllm from https://github.com/vllm-project/vllm")
-
-    warmup = 10
-    rep = 1000
-
-    dtype = torch.float16
-    eps = 1e-5
-    x_shape = (SEQUENCE_TOTAL, HIDDEN_SIZE)
-    w_shape = (x_shape[-1],)
-    residual = torch.rand(x_shape, dtype=dtype, device="cuda")
-    weight = torch.ones(w_shape, dtype=dtype, device="cuda")
-    vllm_norm = RMSNorm(hidden_size=HIDDEN_SIZE, eps=eps).to(dtype=dtype, device="cuda")
-    x = -2.3 + 0.5 * torch.randn(x_shape, dtype=dtype, device="cuda")
-    if provider == "vllm_rms_layernorm":
-        fn = lambda: vllm_norm(x)
-    elif provider == "triton_rms_layernorm":
-        fn = lambda: rms_layernorm(x, weight, eps=eps)
-    elif provider == "vllm_rms_layernorm_with_residual":
-        fn = lambda: vllm_norm(x, residual=residual)
-    elif provider == "triton_rms_layernorm_with_residual":
-        fn = lambda: rms_layernorm(x, weight, eps=eps, residual=residual)
-    else:
-        raise ValueError("Undefined provider.")
-
-    ms = triton.testing.do_bench(fn, warmup=warmup, rep=rep)
-
-    return ms
-
-
 if __name__ == "__main__":
     test_layer_norm()
-    # benchmark_rms_layernorm.run(save_path=".", print_data=True)
